+++ app/src/main/java/com/sopt/now/ApiFactory.kt
@@ -3,9 +3,9 @@ package com.sopt.now
 import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
 import com.sopt.now.BuildConfig.AUTH_BASE_URL
 import com.sopt.now.BuildConfig.OPEN_BASE_URL
-import com.sopt.now.service.AuthService
-import com.sopt.now.service.FriendService
-import com.sopt.now.service.UserService
+import com.sopt.now.data.service.AuthService
+import com.sopt.now.data.service.FriendService
+import com.sopt.now.data.service.UserService
 import kotlinx.serialization.json.Json
 import okhttp3.MediaType.Companion.toMediaType
 import retrofit2.Retrofit
