--- app/src/main/java/com/sopt/now/SignUpState.kt
@@ -1,6 +1,0 @@
-package com.sopt.now
-
-data class SignUpState(
-    val isSuccess: Boolean,
-    val message: String
-)
