--- app/src/main/java/com/sopt/now/activity/SignUpActivity.kt
@@ -1,99 +1,0 @@
-package com.sopt.now.activity
-
-import android.content.Intent
-import android.os.Bundle
-import android.util.Log
-import android.widget.Toast
-import androidx.activity.viewModels
-import androidx.appcompat.app.AppCompatActivity
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
-import com.sopt.now.ServicePool
-import com.sopt.now.SignUpViewModel
-import com.sopt.now.databinding.ActivitySignUpBinding
-import com.sopt.now.request.RequestSignUpDto
-import com.sopt.now.response.ResponseSignUpDto
-
-class SignUpActivity : AppCompatActivity() {
-
-    private val binding by lazy { ActivitySignUpBinding.inflate(layoutInflater) }
-    private val viewModel by viewModels<SignUpViewModel>()
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(binding.root)
-        initViews()
-        initObserver()
-    }
-
-    private fun initViews() {
-        binding.btnSignUp.setOnClickListener {
-            viewModel.signUp(getSignUpRequestDto())
-        }
-    }
-
-    private fun initObserver() {
-        viewModel.liveData.observe(this) {
-            Toast.makeText(
-                this@SignUpActivity,
-                it.message,
-                Toast.LENGTH_SHORT,
-            ).show()
-        }
-    }
-
-//    private fun signUp() {
-//        val signUpRequest = getSignUpRequestDto()
-//        authService.signUp(signUpRequest).enqueue(object : Callback<ResponseSignUpDto> {
-//            override fun onResponse(
-//                call: Call<ResponseSignUpDto>,
-//                response: Response<ResponseSignUpDto>,
-//            ) {
-//                if (response.isSuccessful) {
-//                    val data: ResponseSignUpDto? = response.body()
-//                    val userId = response.headers()["location"]
-//                    Toast.makeText(
-//                        this@SignUpActivity,
-//                        "회원가입 성공 유저의 ID는 $userId 입니둥",
-//                        Toast.LENGTH_SHORT,
-//                    ).show()
-//                    Log.d("SignUp", "data: $data, userId: $userId")
-//                    navigateToLogin(userId)
-//                } else {
-//                    val error = response.message()
-//                    Toast.makeText(
-//                        this@SignUpActivity,
-//                        "로그인이 실패 $error",
-//                        Toast.LENGTH_SHORT,
-//                    ).show()
-//                }
-//            }
-//
-//            override fun onFailure(call: Call<ResponseSignUpDto>, t: Throwable) {
-//                Toast.makeText(this@SignUpActivity, "서버 에러 발생 ", Toast.LENGTH_SHORT).show()
-//            }
-//        })
-//    }
-
-    private fun navigateToLogin(userId: String?) {
-        val intent = Intent(this, LoginActivity::class.java).apply {
-            putExtra("userId", userId)
-        }
-        startActivity(intent)
-        finish()
-    }
-
-    private fun getSignUpRequestDto(): RequestSignUpDto {
-        val id = binding.etSignUpId.text.toString()
-        val password = binding.etSignUpPwd.text.toString()
-        val nickname = binding.etSignUpNickname.text.toString()
-        val phoneNumber = binding.etSignUpPhone.text.toString()
-        return RequestSignUpDto(
-            authenticationId = id,
-            password = password,
-            nickname = nickname,
-            phone = phoneNumber
-        )
-    }
-}
