+++ app/src/main/java/com/sopt/now/data/dto/request/RequestSignUpDto.kt
@@ -1,0 +1,16 @@
+package com.sopt.now.data.dto.request
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RequestSignUpDto(
+    @SerialName("authenticationId")
+    val authenticationId: String,
+    @SerialName("password")
+    val password: String,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("phone")
+    val phone: String,
+)
