+++ app/src/main/java/com/sopt/now/presentation/home/HomeViewModel.kt
@@ -1,0 +1,72 @@
+package com.sopt.now.presentation.home
+
+import androidx.lifecycle.ViewModel
+import com.sopt.now.presentation.friend.Friend
+
+class HomeViewModel : ViewModel() {
+
+    val mockFriendList = mutableListOf<Friend>()
+
+//    val mockFriendList = listOf<Friend>(
+//        Friend(
+//            profileImage = R.drawable.carrot,
+//            name = "jihyun",
+//            selfDescription = "안녕하세용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.heart,
+//            name = "gaeul",
+//            selfDescription = "저희 2조는",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.soccer,
+//            name = "seokchan",
+//            selfDescription = "석류톡톡이에용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.carrot,
+//            name = "jihyun",
+//            selfDescription = "안녕하세용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.heart,
+//            name = "gaeul",
+//            selfDescription = "저희 2조는",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.soccer,
+//            name = "seokchan",
+//            selfDescription = "석류톡톡이에용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.carrot,
+//            name = "jihyun",
+//            selfDescription = "안녕하세용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.heart,
+//            name = "gaeul",
+//            selfDescription = "저희 2조는",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.soccer,
+//            name = "seokchan",
+//            selfDescription = "석류톡톡이에용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.carrot,
+//            name = "jihyun",
+//            selfDescription = "안녕하세용",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.heart,
+//            name = "gaeul",
+//            selfDescription = "저희 2조는",
+//        ),
+//        Friend(
+//            profileImage = R.drawable.soccer,
+//            name = "seokchan",
+//            selfDescription = "석류톡톡이에용",
+//        )
+//    )
+}
