+++ app/src/main/java/com/sopt/now/presentation/main/MainActivity.kt
@@ -1,0 +1,64 @@
+package com.sopt.now.presentation.main
+
+import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
+import androidx.fragment.app.Fragment
+import com.sopt.now.R
+import com.sopt.now.databinding.ActivityMainBinding
+import com.sopt.now.presentation.home.HomeFragment
+import com.sopt.now.presentation.mypage.MyPageFragment
+import com.sopt.now.presentation.search.SearchFragment
+
+
+class MainActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityMainBinding
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        binding = ActivityMainBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        initHomeFragment()
+        clickBottomNavigation()
+    }
+
+    private fun initHomeFragment(){
+        val currentFragment = supportFragmentManager.findFragmentById(binding.fcvHome.id)
+        if (currentFragment == null) {
+            supportFragmentManager.beginTransaction()
+                .add(binding.fcvHome.id, HomeFragment())
+                .commit()
+        }
+    }
+
+    private fun clickBottomNavigation() {
+        binding.bnvHome.setOnItemSelectedListener{
+            when (it.itemId) {
+                R.id.menu_home-> {
+                    replaceFragment(HomeFragment())
+                    true
+                }
+
+                R.id.menu_search-> {
+                    replaceFragment(SearchFragment())
+                    true
+                }
+
+                R.id.menu_mypage-> {
+                    replaceFragment(MyPageFragment())
+                    true
+                }
+
+                else -> false
+            }
+        }
+    }
+
+    private fun replaceFragment(fragment: Fragment) {
+        supportFragmentManager.beginTransaction()
+            .replace(R.id.fcv_home, fragment)
+            .commit()
+    }
+
+}
