+++ app/src/main/java/com/sopt/now/presentation/signup/SignUpState.kt
@@ -1,0 +1,6 @@
+package com.sopt.now.presentation.signup
+
+data class SignUpState(
+    val isSuccess: Boolean,
+    val message: String
+)
