--- app/src/main/java/com/sopt/now/request/RequestLoginDto.kt
@@ -1,12 +1,0 @@
-package com.sopt.now.request
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RequestLoginDto(
-    @SerialName("authenticationId")
-    val authenticationId: String,
-    @SerialName("password")
-    val password: String
-)
