--- app/src/main/java/com/sopt/now/response/ResponseFriendDto.kt
@@ -1,29 +1,0 @@
-package com.sopt.now.response
-
-import com.sopt.now.friend.Friend
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseFriendDto(
-    @SerialName("page")
-    val page: Int,
-    @SerialName("per_page")
-    val perPage: Int,
-    @SerialName("total")
-    val total: Int,
-    @SerialName("total_pages")
-    val totalPages: Int,
-    @SerialName("data")
-    val data: List<Friend>,
-    @SerialName("support")
-    val support: Support
-) {
-    @Serializable
-    data class Support(
-        @SerialName("url")
-        val url: String,
-        @SerialName("text")
-        val text: String
-    )
-}
