--- app/src/main/java/com/sopt/now/response/ResponseSignUpDto.kt
@@ -1,12 +1,0 @@
-package com.sopt.now.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ResponseSignUpDto(
-    @SerialName("code")
-    val code: Int,
-    @SerialName("message")
-    val message: String,
-)
