--- app/src/main/java/com/sopt/now/search/SearchFragment.kt
@@ -1,13 +1,0 @@
-package com.sopt.now.search
-
-import android.os.Bundle
-import android.view.View
-import com.sopt.now.BindingFragment
-import com.sopt.now.databinding.FragmentSearchBinding
-
-class SearchFragment : BindingFragment<FragmentSearchBinding>(FragmentSearchBinding::inflate) {
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-    }
-}
