--- app/src/main/java/com/sopt/now/user/UserInfo.kt
@@ -1,10 +1,0 @@
-package com.sopt.now.user
-
-import android.os.Parcelable
-import kotlinx.parcelize.Parcelize
-
-data class UserInfo(
-    val authenticationId: String,
-    val nickname: String,
-    val phone: String
-)
